refactor: cleanup providers for LLM to prep for future ops
